pyproject.toml
[tool.poetry]
name = "ai-real-estate-advisor"
version = "0.1.0"
description = "AI Real Estate Advisor App"
authors = ["Asif Shah <asif.shah9091@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.0,<3.13"
openai = "^1.101.0"
streamlit = "^1.48.1"
# ImportError: libcudnn.so.9: cannot open shared object file: No such file or directory
langchain-community = "^0.3.27"
python-dotenv = "^1.1.1"
langchain-experimental = "^0.3.4"
langchain-openai = "^0.3.31"
tabulate = "^0.9.0"
pyzipcode = "^3.0.1"
folium = "^0.20.0"
streamlit-folium = "^0.25.1"
faker = "^37.5.3"
duckduckgo_search = "^8.1.1"
langchainhub = "^0.1.21"
langchain-core = "^0.3.74"
langchain-text-splitters = "^0.3.9"
validators = "^0.35.0"
fastembed = "^0.7.1"
chromadb = "^1.0.20"
sentence-transformers = "^5.1.0"
numpy = "^2.3.2"
transformers = "^4.55.4"
onnxruntime = "^1.22.1"
ragas = "^0.3.2"
langsmith = "^0.4.16"
pandas = "^2.3.2"
scikit-learn = "^1.7.1"
scipy = "^1.16.1"
torch = "^2.2.2"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
tqdm = "^4.67.1"
rich = "^14.1.0"
click = "^8.2.1"
typer = "^0.16.1"
httpx = "^0.28.1"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
starlette = "^0.47.2"
anyio = "^4.10.0"
nest-asyncio = "^1.6.0"
uvloop = "^0.21.0"
orjson = "^3.11.2"
pyarrow = "^21.0.0"
safetensors = "^0.6.2"
bcrypt = "^4.3.0"
cryptography = "^45.0.6"
rsa = "^4.9.1"
watchfiles = "^1.1.0"
fsspec = "^2025.7.0"
filelock = "^3.19.1"
websockets = "^15.0.1"
websocket-client = "^1.8.0"
aiohttp = "^3.12.15"
SQLAlchemy = "^2.0.43"
diskcache = "^5.6.3"
python-dateutil = "^2.9.0.post0"
pytz = "^2025.2"
tzdata = "^2025.2"
regex = "^2025.7.34"
markdown-it-py = "^4.0.0"
Pygments = "^2.19.2"
sympy = "^1.14.0"
mpmath = "^1.3.0"
build = "^1.3.0"
setuptools = "^80.9.0"
wheel = "^0.45.1"
mypy-extensions = "^1.1.0"
typing-extensions = "^4.14.1"
typing-inspect = "^0.9.0"
typing-inspection = "^0.4.1"
colorama = "^0.4.6"
coloredlogs = "^15.0.1"
loguru = "^0.7.3"
tenacity = "^9.1.2"
backoff = "^2.2.1"
lxml = "^6.0.1"
cffi = "^1.17.1"
pycparser = "^2.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
